implement move system; tie game loop to "constant" fps value

+++ src/ecs/components/position.h
@@ -3,6 +3,6 @@
 #include "../../types.h"
 
 typedef struct Position {
-    u32 X;
-    u32 Y;
+    f32 X;
+    f32 Y;
 } Position;
